Rename styles from edit to popup
We have a good styles which describe edit-popup. These styles can be
used for other popup-components. So, let make these styles as a 'popup'
and reuse for a new stuff.
